pyproject.toml
[project]
name = "py-kobotoolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.32",
    "python-dotenv>=1.0",
]
